+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * Add GAE ENV_VARs to trace context when starting a trace
 ### Removed
 ### Changed
 ### Fixed